==> project/code/00_DoNotRun.ipynb <==
# Preprocessing

### Demonstration Code Disclaimer

Please note that the code included in this document is intended **solely for demonstration purposes** and is not designed to be executed in its current form. It was originally crafted to process extensive datasets, which due to their size, could not be hosted on GitHub.

### Data Source

The data consists of two CSV files that have been acquired from [Kaggle](SOURCE).

### Python Notebook Objective

The primary focus of this Python notebook is to:

1. Filter the dataset to include only players from specific clubs.
2. Merge the tables into a single dataset.
3. Compute new features based on the combined data.

## Import libraries, define the working directory and load the data

import pandas as pd
# Define the path to the data folder
data_folder = 'data/original/'

To work with the data, we first load the CSV files into Pandas DataFrames using the following code snippet:

# Load appearances data
appearances_df = pd.read_csv(data_folder + 'appearances.csv')

# Load clubs data
clubs_df = pd.read_csv(data_folder + 'clubs.csv')

# Load players data
players_df = pd.read_csv(data_folder + 'players.csv')


## Players data 

This section outlines the steps taken to filter the player data based on selected competitions and to save the filtered dataset.



### Displaying Initial Data

Firstly, we display the first few entries of the player data to understand its structure:

# Display the first few rows of the players DataFrame
players_df.head()

### Filtering 
Here, we define the list of competition IDs that we are interested in

# Define the list of competition IDs to include
competition_ids = ['GB1', 'IT1', 'ES1', 'L1', 'FR1']

Next, we filter the players_df DataFrame to include only the players from the specified competition

# Filter the players DataFrame by the selected competition IDs
players_df_filtered = players_df[players_df['current_club_domestic_competition_id'].isin(competition_ids)]

### Selecting Relevant Columns
We then specify the columns we want to retain in our filtered DataFrame:

# Define the columns to keep
columns_players = [
    'name',
    'current_club_id',
    'country_of_citizenship',
    'date_of_birth',
    'position',
    'foot', 
    'height_in_cm',
    'market_value_in_eur',
    'highest_market_value_in_eur',
    'player_id'
]

# Apply the column filter to the DataFrame
players_df_filtered = players_df_filtered[columns_players]

### Display Filtered Data
After filtering, we can display the resulting DataFrame:

players_df_filtered

### Saving the Filtered Data
Finally, we save the filtered data to a CSV file:

# Save the filtered DataFrame to a CSV file
players_df_filtered.to_csv('data/filtered/players_filtered.csv')

## Appearances data
This section of the notebook focuses on processing the appearances data for a set of football players. We'll filter the data to include only relevant statistics for each player and then save this filtered dataset for future analysis.

### Define Columns of Interest

We will filter the DataFrame to only include the columns that are relevant to our analysis.

# Define the columns that we are interested in
columns_appearances = [
    'player_id',
    'yellow_cards',
    'red_cards',
    'goals',
    'assists',
    'minutes_played'
]

# Filter the DataFrame to include only the columns of interest
appearances_df_filtered = appearances_df[columns_appearances]

### Filter Data Based on Unique Player IDs

We will further filter the data based on a list of unique player IDs to ensure we're analyzing the correct subset of players.

# Assuming players_df_filtered has been defined previously and contains the 'player_id' column
player_ids = players_df_filtered['player_id'].unique()

# Filter the appearances DataFrame to only include rows with player IDs that are in our list of unique player IDs
appearances_df_filtered = appearances_df[appearances_df['player_id'].isin(player_ids)]

appearances_df.head()

### Save the Filtered Data

Finally, we will save the filtered data to a CSV file for further analysis or future use.

# Save the filtered DataFrame to a CSV file
appearances_df_filtered.to_csv('data/filtered/appearances_filtered.csv')
